+++ src/main/java/org/ibd/model/weapons/Pistol.java
@@ -2,6 +2,8 @@ package org.ibd.model.weapons;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @DiscriminatorValue("pIsToL")
 @Access(AccessType.FIELD)
@@ -9,7 +11,7 @@ public class Pistol extends Firearm {
     public Pistol() {
     }
 
-    public Pistol(Long serialNumber, String manufacturer, String name, Float price, String caliber) {
+    public Pistol(Long serialNumber, String manufacturer, String name, BigDecimal price, String caliber) {
         super(serialNumber, manufacturer, name, price, caliber);
         setType("Pistol");
     }
