+++ src/main/java/org/ibd/model/weapons/RecreationalMcNuke.java
@@ -2,6 +2,8 @@ package org.ibd.model.weapons;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @DiscriminatorValue("McNuke:D")
 @Access(AccessType.FIELD)
@@ -10,7 +12,7 @@ public class RecreationalMcNuke extends Explosive {
     }
 
 
-    public RecreationalMcNuke(Long serialNumber, String manufacturer, String name, Float price, Integer power) {
+    public RecreationalMcNuke(Long serialNumber, String manufacturer, String name, BigDecimal price, Integer power) {
         super(serialNumber, manufacturer, name, price, power);
         setType("Nuke");
